requirements.txt
Flask==3.1.1
openai==1.97.0
requests==2.32.4
python-dotenv==1.0.0
azure-cosmos==4.6.0
azure-identity==1.17.0
